fastApi/requirements.txt
fastapi==0.109.0
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.12.0
uvicorn==0.27.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
aiohttp==3.9.3 